backendfastapi/requirements.txt
fastapi
uvicorn
tensorflow
tensorflowjs
numpy
google-cloud-firestore
google-cloud-storage
requests
pytest

